% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{summary.oneway}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{x}{An object of class "oneway" and "list"
containing the results of a oneway analysis of variance
as well as summary statistics from the data used}

\item{...}{additional arguments that may be passed to print()}
}
\value{
returns an object of class "anova"
and "data.frame"
}
\description{
Creates a summary of an
object of class "oneway"
}
\examples{
results <- oneway(mpg ~ cyl, mtcars)
summary.oneway(results)
}
